+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - test command for testing with bats
 - prevent double root dir
+- use custom test command
 ### Changed
 - Indicate progress while downloading
 - Lock current dirs to actual OS/ARCH